pyproject.toml
[tool.poetry]
name = "cadastr-api-test-case"
version = "0.1.0"
description = ""
authors = ["Andrey Barinov <barbpro89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
djangorestframework = "^3.15.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
